refactor participant list, filter route
