--- src/main/java/org/broadinstitute/dsm/model/filter/participant/FilterFactory.java
@@ -1,4 +1,0 @@
-package org.broadinstitute.dsm.model.filter.participant;
-
-public class FilterFactory {
-}
