--- src/main/java/org/broadinstitute/dsm/model/filter/tissue/DefaultFilterTissueList.java
@@ -1,58 +1,0 @@
-package org.broadinstitute.dsm.model.filter.tissue;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Objects;
-
-import org.apache.commons.lang3.StringUtils;
-import org.broadinstitute.dsm.db.DDPInstance;
-import org.broadinstitute.dsm.db.ViewFilter;
-import org.broadinstitute.dsm.model.TissueList;
-import org.broadinstitute.dsm.model.TissueListWrapper;
-import org.broadinstitute.dsm.model.filter.Filterable;
-import org.broadinstitute.dsm.route.FilterRoute;
-import org.broadinstitute.dsm.statics.RequestParameter;
-import org.broadinstitute.dsm.statics.RoutePath;
-import org.broadinstitute.dsm.util.PatchUtil;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import spark.QueryParamsMap;
-
-public class DefaultFilterTissueList extends BaseFilterTissueList {
-
-    private static final Logger logger = LoggerFactory.getLogger(DefaultFilterTissueList.class);
-
-    public DefaultFilterTissueList(String jsonBody) {
-        super(jsonBody);
-    }
-
-    @Override
-    public List<TissueListWrapper> filter(QueryParamsMap queryParamsMap) {
-        List<TissueListWrapper> wrapperList = new ArrayList<>();
-        if (Objects.isNull(queryParamsMap)) return wrapperList;
-        prepareNeccesaryData(jsonBody, queryParamsMap);
-        if (!queryParamsMap.hasKey(RoutePath.REALM)) throw new NoSuchElementException("realm is necessary");
-        String realm = queryParamsMap.get(RoutePath.REALM).value();
-        DDPInstance ddpInstance = DDPInstance.getDDPInstance(realm);
-        String defaultFilter = queryParamsMap.get(RoutePath.FILTER_DEFAULT).value();
-        List<TissueList> tissueListList;
-        if ("0".equals(defaultFilter)) {
-            tissueListList = TissueList.getAllTissueListsForRealmNoFilter(realm);
-            wrapperList = TissueListWrapper.getTissueListData(ddpInstance, null, tissueListList);
-        } else if ("1".equals(defaultFilter)) {
-            String userEmail = queryParamsMap.value(RequestParameter.USER_MAIL);
-            String defaultFilterName = ViewFilter.getDefaultFilterForUser(userEmail, "tissueList");
-            if (StringUtils.isNotBlank(defaultFilterName)) {
-                wrapperList = (List<TissueListWrapper>) FilterRoute.getListBasedOnFilterName(defaultFilterName, realm, "tissueList", null, null);
-            } else {
-                tissueListList = TissueList.getAllTissueListsForRealmNoFilter(realm);
-                wrapperList = TissueListWrapper.getTissueListData(ddpInstance, null, tissueListList);
-            }
-        } else {
-            wrapperList = filterTissueList(filters, PatchUtil.getColumnNameMap(), quickFilterName, ddpInstance, filterQuery);
-        }
-        logger.info("Found " + wrapperList.size() + " tissues for Tissue View");
-        return wrapperList;
-    }
-}
